hello-next-js refactor: tidy up TASK CRUD - part 3
isSafeInput fn is reusable
