+++ myapps/hello-next-js/src/components/Task/use-client/TaskTable.tsx
@@ -5,7 +5,7 @@
 import { useState, useCallback, Dispatch, SetStateAction } from 'react';
 import { useRouter } from 'next/router';
 import type { Task } from "@/types/Task";
-import { MONOREPO_PREFIX, TASKS_CRUD } from "@/lib/app/common";
+import { MONOREPO_PREFIX, TASKS_CRUD, isSafeInput } from "@/lib/app/common";
 
 export type TaskTableType = {
     tasks: Task[], 
@@ -15,33 +15,6 @@ export type TaskTableType = {
     setButtonDisabled: Dispatch<SetStateAction<boolean>>,
     userAuthenticated: boolean,
 }
-
-const isSafeInput = (str: string) => {
-    // for reference: To prevent SQL injection attack
-    // Only allow alphanumeric characters, basic punctuation, and whitespace
-    //const regex = /^[a-zA-Z0-9\s.,!?'"()\-_:;]{1,500}$/;
-    //return regex.test(str);
-
-    // Length check
-    if (str.length === 0 || str.length > 500) return false;
-
-    // Character whitelist - removed potentially dangerous chars
-    const allowedChars = /^[a-zA-Z0-9\s.,!?()\-_:]+$/;
-    if (!allowedChars.test(str)) return false;
-
-    // Blacklist dangerous patterns
-    const dangerousPatterns = [
-        /--;/,                    // SQL comment
-        /\/\*/,                   // Multi-line comment start
-        /\*\//,                   // Multi-line comment end
-        /<script/i,               // Script tag
-        /javascript:/i,           // JavaScript protocol
-        /on\w+\s*=/i,            // Event handlers
-        /drop|delete|insert|update|select|union|exec/i // SQL keywords
-    ];
-
-    return !dangerousPatterns.some(pattern => pattern.test(str));
-};
 
 export const TaskTable = ({ tasks, createRow, updateRowFromId, buttonDisabled, setButtonDisabled, userAuthenticated } : TaskTableType) => {
     const [title, setTitle] = useState("");
