+++ myapps/hello-next-js/src/components/Task/use-server/TaskTable.tsx
@@ -7,7 +7,7 @@
 import { useCallback, useRef, Dispatch, SetStateAction } from 'react';
 import { useRouter, usePathname } from 'next/navigation';
 import type { Task } from "@/types/Task";
-import { MONOREPO_PREFIX, TASKS_CRUD } from "@/lib/app/common";
+import { MONOREPO_PREFIX, TASKS_CRUD, isSafeInput } from "@/lib/app/common";
 
 // for reference: 
 // ** ->: the viewmodel fn returns the promise of updated Tasks, not 
@@ -22,33 +22,6 @@ type TaskTableDefaultType = {
     userAuthenticated: boolean,
 }
 export type TaskTableType = TaskTableDefaultType;
-
-const isSafeInput = (str: string) => {
-    // for reference: To prevent SQL injection attack
-    // Only allow alphanumeric characters, basic punctuation, and whitespace
-    //const regex = /^[a-zA-Z0-9\s.,!?'"()\-_:;]{1,500}$/;
-    //return regex.test(str);
-
-    // Length check
-    if (str.length === 0 || str.length > 500) return false;
-
-    // Character whitelist - removed potentially dangerous chars
-    const allowedChars = /^[a-zA-Z0-9\s.,!?()\-_:]+$/;
-    if (!allowedChars.test(str)) return false;
-
-    // Blacklist dangerous patterns
-    const dangerousPatterns = [
-        /--;/,                    // SQL comment
-        /\/\*/,                   // Multi-line comment start
-        /\*\//,                   // Multi-line comment end
-        /<script/i,               // Script tag
-        /javascript:/i,           // JavaScript protocol
-        /on\w+\s*=/i,            // Event handlers
-        /drop|delete|insert|update|select|union|exec/i // SQL keywords
-    ];
-
-    return !dangerousPatterns.some(pattern => pattern.test(str));
-};
 
 export const TaskTable = ({ tasks, setTasks, createRow, updateRowFromId, buttonDisabled, setButtonDisabled, userAuthenticated } : TaskTableType) => {
     const appRouter = useRouter();
