+++ myapps/hello-next-js/src/components/Task/use-server/TaskTableWithSwr.test.tsx
@@ -4,6 +4,7 @@ import '@testing-library/jest-dom';
 import { usePathname, useRouter } from 'next/navigation';
 import { Task } from "@/types/Task";
 import { TaskTableWithSwr, TaskTableType } from './TaskTableWithSwr';
+import { isSafeInput } from '@/lib/app/common';
 
 // mock the http only auth_token cookie. 
 // The presence of this cookie indicates that the user has logged in
@@ -28,6 +29,7 @@ jest.mock('next/navigation', () => ({
 jest.mock("../../../lib/app/common", () => ({
     MONOREPO_PREFIX: '/app',
     TASKS_CRUD: 'tasks',
+    isSafeInput: jest.fn().mockReturnValue(true),
 }));
 
 let spyConsoleError: jest.SpyInstance<any, any>;
@@ -64,8 +66,11 @@ describe('TaskTableWithSwr Component', () => {
 
     afterEach(() => {
         spyConsoleError.mockRestore();
-        jest.clearAllMocks();
     });
+
+    afterAll(() => {
+        jest.restoreAllMocks();
+    })
 
     const defaultProps: TaskTableType = {
         tasks: mockTasks,
@@ -191,6 +196,8 @@ describe('TaskTableWithSwr Component', () => {
         });
 
         it('rejects input with HTML tags', async () => {
+            (isSafeInput as jest.Mock).mockReturnValue(false);
+                        
             render(<TaskTableWithSwr {...defaultProps} />);
             
             const titleInput = screen.getByPlaceholderText('Title');
@@ -207,6 +214,8 @@ describe('TaskTableWithSwr Component', () => {
 
     describe('Edge cases', () => {
         it('rejects input that is too long', async () => {
+            (isSafeInput as jest.Mock).mockReturnValue(false);
+                        
             render(<TaskTableWithSwr {...defaultProps} />);
             
             const titleInput = screen.getByPlaceholderText('Title');
