+++ myapps/hello-next-js/src/lib/app/common.ts
@@ -165,3 +165,30 @@ export const VERIFY_JWT_RETURN_API_RES = async (req: NextApiRequest, res: NextAp
 
     return true;
 }
+
+export const isSafeInput = (str: string) => {
+    // for reference: To prevent SQL injection attack
+    // Only allow alphanumeric characters, basic punctuation, and whitespace
+    //const regex = /^[a-zA-Z0-9\s.,!?'"()\-_:;]{1,500}$/;
+    //return regex.test(str);
+
+    // Length check
+    if (str.length === 0 || str.length > 500) return false;
+
+    // Character whitelist - removed potentially dangerous chars
+    const allowedChars = /^[a-zA-Z0-9\s.,!?()\-_:]+$/;
+    if (!allowedChars.test(str)) return false;
+
+    // Blacklist dangerous patterns
+    const dangerousPatterns = [
+        /--;/,                    // SQL comment
+        /\/\*/,                   // Multi-line comment start
+        /\*\//,                   // Multi-line comment end
+        /<script/i,               // Script tag
+        /javascript:/i,           // JavaScript protocol
+        /on\w+\s*=/i,            // Event handlers
+        /drop|delete|insert|update|select|union|exec/i // SQL keywords
+    ];
+
+    return !dangerousPatterns.some(pattern => pattern.test(str));
+};
